# metal : multi-simd softmax kernel